requirements.txt
pandas
matplotlib

torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1

accelerate
diffusers==0.30.3
einops
lightning
transformers==4.44.2
kornia
open_clip_torch

click
hydra-core
hydra-colorlog
omegaconf
wandb

decord
librosa
opencv-python-headless
soundfile
moviepy

easydict
rootutils
blobfile
